update simsdp project into integration test

+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/node/simulator/RadarExporterTask.java
@@ -244,7 +244,7 @@ public class RadarExporterTask extends AbstractTaskImplementation {
       final String lineToConsider = lines[lines.length - 1];
       final String[] memPerNode = lineToConsider.split(";");
       for (final String element : memPerNode) {
-        memory += Integer.parseInt(element);
+        memory += Long.parseLong(element);
       }
 
     }
